+++ examples/retry.cpp
@@ -21,7 +21,7 @@
 #include <iostream>
 #include <nRF24L01p/nrf24l01p.h>
 
-#include "../util/time_util.h"
+#include "../time_util.h"
 
 using namespace std;
 
@@ -74,8 +74,8 @@ int main(int argc, char** argv) {
         unsigned long packets_failed = 0;
         unsigned long packets_sent = 0;
 
-        long long start = TimeUtil::current_timestamp_milliseconds();
-        long long now = start;
+        unsigned long start = TimeUtil::current_timestamp_milliseconds();
+        unsigned long now = start;
 
         // Put transceiver in standby2 mode for fast transmitting
         nRF24L01p.set_standby2();
@@ -102,7 +102,7 @@ int main(int argc, char** argv) {
             packets_sent++;
 
             now = TimeUtil::current_timestamp_milliseconds();
-            long long elapsed_milliseconds = now - start;
+            unsigned long elapsed_milliseconds = now - start;
 
             if (elapsed_milliseconds > 1000) {
                 float bytes_received = packets_sent * 32;
@@ -128,12 +128,12 @@ int main(int argc, char** argv) {
 
         unsigned long received_packets = 0;
 
-        long long start = TimeUtil::current_timestamp_milliseconds();
-        long long now = start;
+        unsigned long start = TimeUtil::current_timestamp_milliseconds();
+        unsigned long now = start;
 
         while (1) {
             now = TimeUtil::current_timestamp_milliseconds();
-            long long elapsed_milliseconds = now - start;
+            unsigned long elapsed_milliseconds = now - start;
             
             // Check for payloads in the RX FIFO. This is faster then
             // first checking rx_data_ready().
